Categorized the classes to seperate files
