+++ EntityFrameworkCore.Procedure/MultiResultProp.cs
@@ -1,0 +1,11 @@
+﻿using System.Reflection;
+
+namespace EntityFrameworkCore.Procedure
+{
+    internal class MultiResultProp
+    {
+        public string Name { get; set; }
+        public int Order { get; set; }
+        public PropertyInfo Info { get; set; }
+    }
+}
