+++ EntityFrameworkCore.Procedure/SingleSet.cs
@@ -2,152 +2,26 @@
 using EntityFrameworkCore.Procedure.Extensions;
 using Microsoft.EntityFrameworkCore;
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Data.Common;
 using System.Data.SqlClient;
-using System.Linq;
-using System.Reflection;
 
 namespace EntityFrameworkCore.Procedure
 {
     /// <summary>
-    /// Extended Type to store List Collection
+    /// SingleSet it reads only first result set of the Procedure,
+    /// SQL Parameter List can be passed to execute the Procedure.
     /// </summary>
-    /// <typeparam name="T"></typeparam>
-    public class RCollection<T> : List<T> where T : class, new()
-    {
-        public readonly Type type;
-
-        public RCollection()
-        {
-            type = typeof(T);
-        }
-
-        public RCollection(IEnumerable<T> collection) : base(collection)
-        {
-            type = typeof(T);
-        }
-
-        public RCollection(int capacity) : base(capacity)
-        {
-            type = typeof(T);
-        }
-    }
-
-    /// <summary>
-    /// Abstract Class for Poco having Multi Result
-    /// </summary>
-    public abstract class MultiResult
-    {
-        internal List<MultiResultProp> GetProperties()
-        {
-            Type t = this.GetType();
-            List<PropertyInfo> props = t.GetProperties().Where(p => p.PropertyType.FullName.StartsWith("EntityFrameworkCore.Procedure.RCollection")).ToList();
-
-            return null;
-        }
-    }
-
-    internal class MultiResultProp
-    {
-        public string Name { get; set; }
-        public int Order { get; set; }
-        public PropertyInfo Info { get; set; }
-    }
-
-
-    public class SingleSet<Input, T> : SingleSet<T> where T : class, new() where Input : ProcedureParam
-    {
-        public SingleSet(DbContext context, string name, string schema = "dbo") : base(context, name, schema)
-        {
-
-        }
-
-        public T FirstRow(Input input)
-        {
-            return FirstRow(input.ToParam());
-        }
-
-        public ICollection<T> FirstResult(Input input)
-        {
-            return FirstResult(input.ToParam());
-        }
-
-
-    }
-
-    public sealed class MultiSet<Input, T> : MultiSet<T> where T : MultiResult, new() where Input : ProcedureParam
-    {
-        internal Dictionary<int, MultiResultProp> resultSetInfo = new Dictionary<int, MultiResultProp>();
-
-        public MultiSet(DbContext context, string name, string schema = "dbo") : base(context, name, schema)
-        {
-
-        }
-
-        private void InitResultProperty()
-        {
-            //resultSetInfo.Add()
-        }
-
-        public T FirstRow(Input input)
-        {
-            return FirstRow(input.ToParam());
-        }
-
-        public ICollection<T> FirstResult(Input input)
-        {
-            return FirstResult(input.ToParam());
-        }
-
-        public IEnumerable AsEnumerable(Input input)
-        {
-            return AsEnumerable(input.ToParam());
-        }
-
-        public T MultiResult(Input input)
-        {
-            return null;
-        }
-
-        public void MapResultSet<Type>(int order, Func<T, RCollection<Type>> type) where Type : class, new()
-        {
-            if (resultSetOrder.ContainsKey(order))
-            {
-                throw new Exception($"Key already exists");
-            }
-            resultSetOrder.Add(order, typeof(Type));
-        }
-    }
-
-    public class MultiSet<T> : SingleSet<T> where T : MultiResult, new()
-    {
-        protected Dictionary<int, Type> resultSetOrder = new Dictionary<int, Type>();
-        public MultiSet(DbContext context, string name, string schema = "dbo") : base(context, name, schema)
-        {
-
-        }
-
-        public IEnumerable AsEnumerable(params SqlParameter[] parameters)
-        {
-            return null;
-        }
-
-        public void MapResultSet(int order, Type type)
-        {
-            if (resultSetOrder.ContainsKey(order))
-            {
-                throw new Exception($"Key already exists");
-            }
-
-            resultSetOrder.Add(order, type);
-        }
-    }
-
+    /// <typeparam name="T">Poco Class to cast first result set of a Procedure.</typeparam>
     public class SingleSet<T> where T : class, new()
     {
+        /// <summary>
+        /// Procedure Name to be executed in Database.
+        /// </summary>
         public string ProcName { get; private set; } = string.Empty;
+        /// <summary>
+        /// Schema of the Procedure
+        /// </summary>
         public string Schema { get; private set; } = string.Empty;
         public string[] paramList { get; protected set; }
